+++ src/btNode.h
@@ -11,7 +11,7 @@ class btNode: public drawable {
  
 public: // place public functions or variables declarations here
     btNode(); 
-    btNode(int _x, int _y, ofColor _color); 
+    btNode(float _x, float _y, ofColor _color); 
     btNode(point p, int _dim, ofColor _color, bool fill); // constructor - used to initialize an object, if no properties are passed
 
     void draw();
